feat: Use rustix in the unix backend
rustix is a crate that provides an interface to system functions on
Unix-like operating systems. It is a wrapper around Linux syscalls on
Linux and libc on other operating systems. In addition to being safer,
since it uses direct syscalls it reduces instruction counts on Linux.

This commit replaces usages of libc with rustix. In addition to removing
all unsafe code from the unix backend, it should also speed it up a
little on Linux. Note that this bumps the MSRV of this crate to v1.63.
This makes the "io-safety" feature useless, so I've made it an empty
feature.

Signed-off-by: John Nunley <dev@notgull.net>